Integrate weather api and add weather functionality

+++ src/app/core/interceptors/auth.interceptor.ts
@@ -9,15 +9,19 @@ import {
 import { catchError, Observable, throwError } from 'rxjs';
 import { AuthService } from '../services/auth.service';
 import { Router } from '@angular/router';
+import { environment } from 'src/environments/environment';
 
 @Injectable()
 export class AuthInterceptor implements HttpInterceptor {
   constructor(private readonly authService: AuthService, private readonly router: Router) {}
 
   intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
+    const isAppApi = request.url.includes(environment.apiUrl);
+    const isExternalApi = request.url.includes(environment.weather.apiUrl);
+
     const authToken = this.authService.authToken;
 
-    if (authToken) {
+    if (authToken && isAppApi) {
       request = request.clone({
         setHeaders: {
           Authorization: `Bearer ${authToken}`,
@@ -25,9 +29,17 @@ export class AuthInterceptor implements HttpInterceptor {
       });
     }
 
+    if (isExternalApi) {
+      request = request.clone({
+        setParams: {
+          appid: environment.weather.apiKey,
+        },
+      });
+    }
+
     return next.handle(request).pipe(
       catchError((error: HttpErrorResponse) => {
-        if (error.status === 403) {
+        if (error.status === 403 && isAppApi) {
           this.authService.removeAuthToken();
           this.router.navigate(['/']);
         }
